add GoogleService-Info.plist to gitignore
I dno't know if this will work since I added the plist file before adding it into gitignore. But even if it did not work it is still not the end of the world.

Just make sure to tell the next developer to set up their own Firebase service if possible